+++ skrl/envs/wrappers/warp/__init__.py
@@ -8,6 +8,7 @@ from skrl import logger
 from skrl.envs.wrappers.warp.base import MultiAgentEnvWrapper, Wrapper
 from skrl.envs.wrappers.warp.gymnasium_envs import GymnasiumWrapper
 from skrl.envs.wrappers.warp.isaaclab_envs import IsaacLabMultiAgentWrapper, IsaacLabWrapper
+from skrl.envs.wrappers.warp.mani_skill_envs import ManiSkillWrapper
 
 
 __all__ = ["wrap_env", "Wrapper", "MultiAgentEnvWrapper"]
@@ -24,6 +25,7 @@ def wrap_env(
         "isaaclab",
         "isaaclab-single-agent",
         "isaaclab-multi-agent",
+        "mani-skill",
         "pettingzoo",
     ] = "auto",
     verbose: bool = True,
@@ -55,6 +57,8 @@ def wrap_env(
                 - ``"brax"``
             * - Isaac Lab
                 - ``"isaaclab"`` (``"isaaclab-single-agent"``)
+            * - ManiSkill
+              - ``"mani-skill"``
 
         .. list-table:: Multi-agent environments |br|
             :header-rows: 1
@@ -95,6 +99,8 @@ def wrap_env(
 
         if _in(["omni.isaac.lab.*", "isaaclab.*"], base_classes):
             return "isaaclab-*"
+        elif _in("mani_skill.envs..*", base_classes):
+            return "mani-skill"
         elif _in("brax.envs..*", base_classes):
             return "brax"
         elif _in("dm_env..*", base_classes):
@@ -126,6 +132,10 @@ def wrap_env(
         if verbose:
             logger.info("Environment wrapper: Brax")
         return BraxWrapper(env)
+    elif wrapper == "mani-skill":
+        if verbose:
+            logger.info("Environment wrapper: ManiSkill")
+        return ManiSkillWrapper(env)
     elif type(wrapper) is str and wrapper.startswith("isaaclab"):
         # use specified wrapper
         if wrapper == "isaaclab-single-agent":